backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic-settings==2.7.0
python-multipart==0.0.20
pillow==11.0.0
celery[redis]==5.4.0
redis==5.2.1
aiofiles==24.1.0
